+++ _stuff/read_grobid.R
@@ -400,7 +400,7 @@ get_refs <- function(xml) {
 
   if (length(refs) > 0) {
     ref_table <- data.frame(
-      bib_id = xml2::xml_attr(refs, "id")
+      ref_id = xml2::xml_attr(refs, "id")
     )
     # ref_table$doi <- xml2::xml_find_first(refs, ".//analytic //idno[@type='DOI']") |>
     #   xml2::xml_text()
@@ -420,7 +420,7 @@ get_refs <- function(xml) {
       sapply(paste, collapse = ", ")
   } else {
     ref_table <- data.frame(
-      bib_id = character(0),
+      ref_id = character(0),
       doi = character(0),
       ref = character(0)
     )
@@ -442,30 +442,30 @@ get_refs <- function(xml) {
     matches <- gregexpr("(?<=ref type=\"bibr\" target=\"#)b\\d+",
                         textrefp$text, perl = TRUE) |>
       regmatches(textrefp$text, m = _)
-    # textrefp$bib_id <- sapply(matches, paste, collapse = ";")
+    # textrefp$ref_id <- sapply(matches, paste, collapse = ";")
     no_targets <- gregexpr("(?<=ref type=\"bibr\">)[^</ref>]*(?=</ref>)",
                            textrefp$text, perl = TRUE) |>
       regmatches(textrefp$text, m = _) |>
       # only keep non-target refs that might be author names
       lapply(\(x) x[grepl("[a-zA-Z]{2,}", x)])
 
-    textrefp$bib_id <- mapply(c, matches, no_targets, SIMPLIFY = FALSE) |>
+    textrefp$ref_id <- mapply(c, matches, no_targets, SIMPLIFY = FALSE) |>
       sapply(paste, collapse = ";")
 
-    citation_table <- textrefp[textrefp$bib_id != "", ]
+    citation_table <- textrefp[textrefp$ref_id != "", ]
     citation_table$text <- lapply(citation_table$text, xml2::read_html) |>
       sapply(xml2::xml_text)
 
     citation_table <- citation_table |>
-      tidyr::separate_longer_delim("bib_id", delim = ";")
+      tidyr::separate_longer_delim("ref_id", delim = ";")
   } else {
-    citation_table = data.frame(bib_id = character(0),
+    citation_table = data.frame(ref_id = character(0),
                                 text = character(0))
   }
 
   return(list(
     references = ref_table,
-    citations = citation_table[, c("bib_id", "text")]
+    citations = citation_table[, c("ref_id", "text")]
   ))
 }